requirements.txt
# List all python libraries for the lambda
keras
tensorflow-cpu
numpy
matplotlib
Pillow
opencv-python
h5py==2.10.0